folder structure changed each fetch call is now inside the hooks folder unused imports removed
