+++ frontend/src/components/CardDetails.js
@@ -1,7 +1,6 @@
 import Button from 'react-bootstrap/Button';
 import Modal from 'react-bootstrap/Modal';
 import React from 'react';
-import re  from './CardComponent'
 
 function MyVerticallyCenteredModal(props) {
     return (
