+++ frontend/src/components/CommentsModal.js
@@ -1,11 +1,8 @@
-import React, {useEffect, useState} from 'react';
+import React, {useEffect} from 'react';
 import Button from 'react-bootstrap/Button';
 import Modal from 'react-bootstrap/Modal';
-import Post from "./Post";
-import AddPostModal from "./AddPostModal";
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faComment } from '@fortawesome/free-solid-svg-icons';
-import { FaCommentDots } from "react-icons/fa";
+import {FaCommentDots} from "react-icons/fa";
+import useComments from "../hooks/useComments";
 
 
 function MyVerticallyCenteredModal(props) {
@@ -26,15 +23,15 @@ function MyVerticallyCenteredModal(props) {
                 {Object.entries(props.comments).map(([commentId, comment]) => (
                     // <Post key={postId} post={post}></Post>
 
-                    <div key={commentId} className="card mb-3" style={{borderRadius:"25px"}}>
+                    <div key={commentId} className="card mb-3" style={{borderRadius: "25px"}}>
                         <div className="card-body">
                             <p className="card-title">{comment.text}</p>
                             {/*<p className="card-text">Post ID: {comment.post.id}</p>*/}
                         </div>
                     </div>
                 ))}
-                <textarea style={{width: "100%", height: "70px", borderRadius:"25px", padding:"0 15px"}}/>
-                <div style={{ display: 'flex', justifyContent: 'flex-end' }}>
+                <textarea style={{width: "100%", height: "70px", borderRadius: "25px", padding: "0 15px"}}/>
+                <div style={{display: 'flex', justifyContent: 'flex-end'}}>
                     <Button variant="secondary" size="sm">Add </Button>
                 </div>
 
@@ -48,26 +45,7 @@ function MyVerticallyCenteredModal(props) {
 
 function CommentsModal(props) {
     const [modalShow, setModalShow] = React.useState(false);
-
-    const [comments, setComments] = useState([]);
-
-    const fetchData = (id) => {
-        const url = 'http://localhost:8080/comments/' + id;
-        fetch(url)
-            .then(response => {
-                if (!response.ok) {
-                    throw new Error('Network response was not ok');
-                }
-                return response.json();
-            })
-            .then(data => {
-                setComments(data);
-                // console.log(data);
-            })
-            .catch(error => {
-                console.error('Error fetching data:', error);
-            });
-    };
+    const {comments, fetchData} = useComments();
 
     useEffect(() => {
         fetchData(props.postId); // Call fetchData function when the component mounts
@@ -87,15 +65,15 @@ function CommentsModal(props) {
                     alignItems: "center"      // Center the icon vertically
                 }}>
                 {/*<FontAwesomeIcon icon={faComment} style={{ fontSize: "20px" }} /> /!* Use the FontAwesome icon *!/*/}
-                <FaCommentDots style={{ fontSize: "20px" }} />
-                <div style={{ display: "flex", flexDirection: "column", height: "100%", marginLeft:"5px" }}>
-                    <p style={{ margin: 0 }}>{comments.length} Answers</p>
+                <FaCommentDots style={{fontSize: "20px"}}/>
+                <div style={{display: "flex", flexDirection: "column", height: "100%", marginLeft: "5px"}}>
+                    <p style={{margin: 0}}>{comments.length} Answers</p>
                 </div>
 
             </Button>
 
             <MyVerticallyCenteredModal
-                comments = {comments}
+                comments={comments}
                 show={modalShow}
                 onHide={() => setModalShow(false)}
             />
