+++ frontend/src/components/courses/CardComponent.js
@@ -1,7 +1,8 @@
 import Button from 'react-bootstrap/Button';
 import Card from 'react-bootstrap/Card';
 import {useState} from "react";
-import CardDetails from "../../components/CardDetails";
+import CardDetails from "../CardDetails";
+import useSetSubject from "../../hooks/useSetSubject";
 
 
 function replaceProgramNames(str) {
@@ -31,42 +32,17 @@ function replaceProgramNames(str) {
 function CardComponent(props) {
     console.log(props.subject)
     const [clicked, setClicked] = useState(props.subject['isTaken']);
-    const [formData, setFormData] = useState(props.subject.name);
+    const formData = props.subject.name;
 
     const handleClick = () => {
-        if (clicked)
-            setClicked(false);
-        else
-            setClicked(true);
+        setClicked(!clicked);
     };
 
-        const handleSubmit = async (e) => {
-            e.preventDefault();
-
-            try {
-                const response = await fetch('http://localhost:8080/subjects', {
-                    method: 'POST',
-                    headers: {
-                        'Content-Type': 'application/json'
-                    },
-                    body: JSON.stringify(formData)
-                });
-
-                if (response.ok) {
-                    // Request successful, do something with the response
-                    console.log('POST request successful');
-                } else {
-                    // Request failed
-                    console.error('POST request failed');
-                }
-            } catch (error) {
-                console.error('Error:', error);
-            }
-        };
+    const handleSubmit = useSetSubject();
 
     return (
         // style={{ height: '200px', margin: "10px 0"}
-        <Card style={{ height: '180px', margin: "10px 0", backgroundColor: clicked ? '#c9f8c9' : '' }}>
+        <Card style={{height: '180px', margin: "10px 0", backgroundColor: clicked ? '#c9f8c9' : ''}}>
             {/*<Card.Img variant="top" src="holder.js/100px180" />*/}
             <Card.Body>
                 <Card.Title style={{
@@ -78,22 +54,25 @@ function CardComponent(props) {
                 }}>{props.subject['name']}</Card.Title>
                 <Card.Text>
                     {props.subject['code']}
-                    <p style={{ overflow: 'hidden', textOverflow: 'ellipsis', whiteSpace: 'nowrap'}}>{replaceProgramNames(props.subject['studyProgram'])}</p>
+                    <p style={{
+                        overflow: 'hidden',
+                        textOverflow: 'ellipsis',
+                        whiteSpace: 'nowrap'
+                    }}>{replaceProgramNames(props.subject['studyProgram'])}</p>
                 </Card.Text>
                 {/*<Button variant={clicked ? "success" : "secondary"} onClick={handleClick}>Земи предмет!</Button>*/}
                 <CardDetails subject={props.subject}></CardDetails>
 
-                <form onSubmit={handleSubmit} style={{display: "inline-block", marginLeft:"5px"}}>
-                        <input type="text" name="name" value={formData.name}  hidden={true}/>
+                <form onSubmit={(event) => handleSubmit(event, formData)} style={{display: "inline-block", marginLeft: "5px"}}>
+                    <input type="text" name="name" value={formData.name} hidden={true}/>
                     {/*<button type="submit" onClick={handleClick}>Земи предмет!</button>*/}
-                    <Button type="submit" variant={clicked ? 'danger' : 'secondary'}  onClick={handleClick}>{clicked ? 'Отстрани предмет!' : 'Земи предмет!'}</Button>
+                    <Button type="submit" variant={clicked ? 'danger' : 'secondary'}
+                            onClick={handleClick}>{clicked ? 'Отстрани предмет!' : 'Земи предмет!'}</Button>
                     {/*style={{ backgroundColor: clicked ? 'green' : '', color: clicked ? 'white' : '' }}*/}
                 </form>
-
-
             </Card.Body>
         </Card>
-        
+
     );
 }
