+++ frontend/src/components/courses/CardComponent.module.css
