--- frontend/src/componentsNew/Courses/CardComponent.module.css
@@ -1,4 +1,0 @@
-.green-button {
-    background-color: #c9f8c9;
-    color: white;
-}
