+++ frontend/src/pages/RegisterPage.js
@@ -1,61 +1,22 @@
-import {useEffect, useState} from 'react';
 import '../App.css';
-import NavigationBar from "../components/NavigationBar";
-import Home from "../components/Home";
 import Button from 'react-bootstrap/Button';
 import Form from 'react-bootstrap/Form';
-import { brows } from 'react-router-dom';
 import ImagePicker from "../components/ImagePicker";
-import Navigation from "../componentsNew/Navigation/Navigation";
+import Navigation from "../components/navigation/Navigation";
+import useRegisterUser from "../hooks/useRegisterUser";
 
 function RegisterPage() {
-    const [username, setUsername] = useState('');
-    const [email, setEmail] = useState('');
-    const [password, setPassword] = useState('');
-    const [successMessage, setSuccessMessage] = useState('');
-    const [selectedImage, setSelectedImage] = useState('');
 
-    const handleSubmit = async (event) => {
-        event.preventDefault();
-        try {
-            const response = await fetch('http://localhost:8080/register', {
-                method: 'POST',
-                headers: {
-                    'Content-Type': 'application/json'
-                },
-                body: JSON.stringify({
-                    username,
-                    email,
-                    password,
-                    image: selectedImage
-                })
-            });
-
-            console.log(response.text())
-            if (response.ok){
-                console.log('Created account successfully');
-                setSuccessMessage('Account created successfully!'); // Set success message
-                setUsername(''); // Clear form fields
-                setEmail('');
-                setPassword('');
-                setSelectedImage('');
-            }
-            else
-                console.error('Creating account failed');
-
-        } catch (error) {
-            console.error('Error:', error);
-        }
-    };
-
-
+    const {
+        username, setUsername, email, setEmail, password, setPassword, successMessage,
+        setSelectedImage,
+        handleSubmit
+    } = useRegisterUser();
     return (
         <div className="App">
             <Navigation></Navigation>
-
-            <Form style={{padding:"35px"}} onSubmit={handleSubmit}>
+            <Form style={{padding: "35px"}} onSubmit={handleSubmit}>
                 <h2>Create new account!</h2>
-
                 <Form.Group className="mb-3" controlId="formBasicUsername">
                     <Form.Label>Username</Form.Label>
                     <Form.Control
@@ -87,16 +48,16 @@ function RegisterPage() {
                 </Form.Group>
 
                 <Form.Group className="mb-3" controlId="formBasicCheckbox">
-                    <Form.Check type="checkbox" label="Accepts Terms & Conditions" required={true} />
+                    <Form.Check type="checkbox" label="Accepts Terms & Conditions" required={true}/>
                 </Form.Group>
 
                 <ImagePicker setSelectedImage={setSelectedImage} imageUrls={[
                     'https://cdn3d.iconscout.com/3d/premium/thumb/man-5692600-4743369.png',
                     'https://cdn3d.iconscout.com/3d/premium/preview/boy-avatar-6299533-5187865.png?f=webp&h=700',
                     'https://png.pngtree.com/png-clipart/20221207/ourmid/pngtree-fashion-boy-avatar-png-image_6514592.png'
-                ]} />
+                ]}/>
 
-                {successMessage && <p style={{color:"darkgreen"}}>{successMessage}</p>}
+                {successMessage && <p style={{color: "darkgreen"}}>{successMessage}</p>}
 
                 <Button variant="primary" type="submit">
                     Create Account
