--- frontend/src/pages/SelectPage.module.css
@@ -1,3 +1,0 @@
-.subjectCard{
-    width: 30%;
-}
